pyproject.toml
[tool.poetry]
name = "drf-hw"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.24.0"
drf-yasg = "^1.21.7"
setuptools = "^70.0.0"
stripe = "^9.9.0"
celery = "^5.4.0"
redis = "^5.0.5"
eventlet = "^0.36.1"
django-celery-beat = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
